+++ src/modules/files/b.json
@@ -1,4 +1,4 @@
-{
+ {
     "a": 11,
     "b": 22,
     "c": 33
